requirements.txt
gradio
pandas
numpy
scikit-learn
matplotlib
seaborn
Pillow
